refactoring: EventForm component structure of main-app
* EventForm는 너무 큰 컴포넌트이므로 작은 컴포넌트로 쪼갬
* App component하위에 AppBody, AppFooter 등 좀더 구조적인 컴포넌트 트리를 가지도록 수정
* atomic 한 컴포넌트는 단일 파일로 분리
