+++ frontend/main-app/src/components/AppDescription.js
@@ -1,0 +1,11 @@
+import {Typography} from "@material-ui/core";
+import React from "react";
+
+export default function AppDescription() {
+	return (
+		<>
+			<Typography>익명으로 질문할 수 있습니다.</Typography>
+			<Typography>강의 중 궁금한 것들을 편하게 질문하세요.</Typography>
+		</>
+	);
+}
