+++ frontend/main-app/src/components/AppTitle.js
@@ -1,0 +1,10 @@
+import {Typography} from "@material-ui/core";
+import React from "react";
+
+export default function AppTitle() {
+	return (
+		<Typography variant="h3" component="h1">
+			바글바글
+		</Typography>
+	);
+}
