+++ frontend/main-app/src/components/EventCodeInputErrorMessageStyle.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export default styled.div`
+	color: red;
+	height: 2rem;
+`;
